requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
openai>=1.0.0
httpx
aiofiles